Extract TypePath to own file
